+++ src/views/tenant/administration/Tenants.js
@@ -1,106 +1,12 @@
 import React, { useState } from 'react'
 import { useSelector } from 'react-redux'
-import { CButton } from '@coreui/react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import {
-  faCheckCircle,
-  faCog,
-  faEdit,
-  faEllipsisV,
-  faExclamationTriangle,
-} from '@fortawesome/free-solid-svg-icons'
+import { faCog } from '@fortawesome/free-solid-svg-icons'
 import { CippPageList } from 'src/components/layout'
-import { CellTip, CellTipIcon } from 'src/components/tables'
-import { CippActionsOffcanvas } from 'src/components/utilities'
-import { useLazyGenericGetRequestQuery } from 'src/store/api/app'
-import Skeleton from 'react-loading-skeleton'
+import { CellTip } from 'src/components/tables'
 import { TitleButton } from 'src/components/buttons'
 import Portals from 'src/data/portals'
-
-const Offcanvas = (row, rowIndex, formatExtraData) => {
-  const [getTenantDetails, tenantDetails] = useLazyGenericGetRequestQuery()
-  const [ocVisible, setOCVisible] = useState(false)
-
-  function loadOffCanvasDetails(domainName) {
-    setOCVisible(true)
-    getTenantDetails({ path: `api/ListTenantDetails?tenantfilter=${domainName}` })
-  }
-
-  function tenantProperty(tenantDetails, propertyName) {
-    return (
-      <>
-        {tenantDetails.isFetching && <Skeleton count={1} width={150} />}
-        {!tenantDetails.isFetching &&
-          tenantDetails.isSuccess &&
-          (tenantDetails.data[propertyName]?.toString() ?? ' ')}
-      </>
-    )
-  }
-  const actions = Portals.map((portal) => ({
-    icon: <FontAwesomeIcon icon={faCog} className="me-2" />,
-    label: portal.label,
-    external: true,
-    color: 'info',
-    link: portal.url.replace(portal.variable, row[portal.variable]),
-  }))
-  return (
-    <>
-      <CButton size="sm" color="link" onClick={() => loadOffCanvasDetails(row.defaultDomainName)}>
-        <FontAwesomeIcon icon={faEllipsisV} />
-      </CButton>
-      <CippActionsOffcanvas
-        title="Tenant Information"
-        extendedInfo={[
-          {
-            label: 'Display Name',
-            value: tenantProperty(tenantDetails, 'displayName'),
-          },
-          {
-            label: 'Business Phones',
-            value: tenantProperty(tenantDetails, 'businessPhones'),
-          },
-          {
-            label: 'Technical Emails',
-            value: tenantProperty(tenantDetails, 'technicalNotificationMails'),
-          },
-          {
-            label: 'Tenant Type',
-            value: tenantProperty(tenantDetails, 'tenantType'),
-          },
-          {
-            label: 'Created',
-            value: tenantProperty(tenantDetails, 'createdDateTime'),
-          },
-          {
-            label: 'AD Connect Enabled',
-            value: tenantProperty(tenantDetails, 'onPremisesSyncEnabled'),
-          },
-          {
-            label: 'AD Connect Sync',
-            value: tenantProperty(tenantDetails, 'onPremisesLastSyncDateTime'),
-          },
-          {
-            label: 'AD Password Sync',
-            value: tenantProperty(tenantDetails, 'onPremisesLastPasswordSyncDateTime'),
-          },
-        ]}
-        actions={[
-          {
-            icon: <FontAwesomeIcon icon={faEdit} className="me-2" />,
-            label: 'Edit Tenant',
-            link: `/tenant/administration/tenants/Edit?tenantFilter=${row.defaultDomainName}&customerId=${row.customerId}`,
-            color: 'warning',
-          },
-          ...actions,
-        ]}
-        placement="end"
-        visible={ocVisible}
-        id={row.id}
-        hideFunction={() => setOCVisible(false)}
-      />
-    </>
-  )
-}
+import { CippTenantOffcanvasRow } from 'src/components/utilities/CippTenantOffcanvas'
 
 const TenantsList = () => {
   const TenantListSelector = useSelector((state) => state.app.TenantListSelector)
@@ -148,7 +54,7 @@ const TenantsList = () => {
     {
       name: 'Actions',
       center: true,
-      cell: Offcanvas,
+      cell: CippTenantOffcanvasRow,
     },
   ]
   const titleButton = (
